🔄 Update feastfox w/ latest work (probably broke something)
Another day, another commit to the feastfox (feast fox) repository. 🦊
The "latest work" description is about as descriptive as a weather
forecast, but here we are pushing code anyway because deadlines
don't care about our documentation standards.

This update represents the culmination of whatever was deemed
"ready enough" to ship. 📦 While we'd love to claim this is
groundbreaking innovation, it's more likely a collection of
incremental improvements that may or may not work as intended
in production.

The changes included are presumably functional, though we make
no guarantees about edge cases or that one weird configuration
nobody remembers setting up three years ago. 🤷‍♂️ As always,
we've done our best to avoid breaking existing functionality,
but the universe has a sense of humor about these things.

Signed-off-by: Tim Gates <tim.gates@iress.com>